added more factory methods, renamed file
